third_party/blink/web_tests/virtual/view-transition-wide-gamut/external/wpt/css/css-view-transitions/finished-promise-defers-cleanup-expected.txt

This is a testharness.js-based test.
[FAIL] View transition cleanup is deferred until after the frame where `finished` promise resolves
  assert_equals: expected "absolute" but got "fixed"
Harness: the test ran to completion.
